マイページController修正

+++ src/main/java/com/example/demo/controller/MypageController.java
@@ -42,7 +42,7 @@ public class MypageController {
 	}
 
 	@GetMapping("mypage/{id}")
-	public String showMoreMypage(@PathVariable(name = "id") Integer id, Model model) {
+	public String showMoreMypage(@PathVariable(name = "id") Long id, Model model) {
 
 		if (id == 1) {
 			List<History> orderDetail = historyRepository.findByAccountId(myAccount.getId());
@@ -67,7 +67,7 @@ public class MypageController {
 
 	// 編集完了 →　マイページへ
 	@PostMapping("/mypage/update")
-	public String updateMypage(@RequestParam(value = "id", defaultValue = "") Integer id,
+	public String updateMypage(@RequestParam(value = "id", defaultValue = "") Long id,
 			@RequestParam(value = "name", defaultValue = "") String name,
 			@RequestParam(value = "password", defaultValue = "") String password,
 			@RequestParam(value = "email", defaultValue = "") String email,
@@ -91,7 +91,7 @@ public class MypageController {
 
 	// ユーザー詳細
 	@GetMapping("/user/{id}/detail")
-	public String showUserDetail(@PathVariable("id") Integer id, Model model) {
+	public String showUserDetail(@PathVariable("id") Long id, Model model) {
 		Account account = accountRepository.findById(id).orElse(null);
 		//ユーザーの商品一覧を取得
 		List<Item> items = itemRepository.findByAccountId(id);
